refactor: 将 PID 从 Algorithm 移到 Controller，定义为控制器；增加 app_motor 扩展功能，增加控制器 ctrl_motor_base_pid 用于对电机的基础 PID 控制，删除原有 app_motor_base_controller。

+++ Components/Controller/ForwardFeed/ctrl_forward_feed.cc
@@ -6,6 +6,6 @@
 
 using namespace Controller;
 
-double ForwardFeed::update(double current, double target) {
+float ForwardFeed::update(float current, float target) {
 	return current + target;
 }
